requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
psycopg==3.2.3
sqlparse==0.5.3
